Add custom mouse commands (i.e. scheduled & backdated clicks)
Add a custom mouse module, with different types of click based on whether you're
using the zoom mouse, or another mouse mode. Clicks can be frontloaded to occur
on the next "full click", so to speak, or they can be performed at the time a
command was spoken. They can also be performed immediately.

There are some issues with this setup - it depends on the BetterZoomMouse and it
seems that Dragon/Draconity's clock can drift out of sync with Python's. Not
sure why this happens - could it be a growing delay in the audio interface?
Either way, restarting helps.