Add prefix to basic numbers to prevent conflict with repeats
Typed numbers and repeats were ambiguous, because they both boiled down to
atomic numbers. Need to disambiguate them. It's much more common that I'll want
to repeat a command than actually type a number, so shift the additional burden
to typing by adding a prefix, "num"

For example:

    "num one" -> 1
    "whale five" -> wwwww
    "whale num three" -> w3

Also make them greedy to prevent command splitting in a list of digits. This
needs to be a consistent practice across the codebase.