Add context that matches on specific websites.
So for example, you could have some commands that are active when you're on
YouTube, others when you're on Facebook.

Currently the implementation is pretty hacky, in the long run it would be better
to do this with browser APIs. It's only really designed for Firefox.

The implementation is designed to be general across common browsers, but it
doesn't work on Chrome because Talon fails to extract exe/title info from it.

Only works on Windows (at the moment).