Pull common vocab into a central location
Certain commands will have common components, and those components might change,
or require alternatives (like contractions). Create a stopgap solution which
allows these components to be pulled into one location.

In the long-run it may be better to create a better structure for this. Probably
best to wait for newapi first.