Don't catch timeout errors, it hides certain successful calls
Currently, Porthole can send a command successfully but timeout while trying to
receive the reply. We don't want to retry for this kind of error because it
could change the Emacs state twice.

Don't bother catching a full timeout given this - there's no point, we wouldn't
want to retry.

Long-term it might be better to add a nonce or handshake to Voicemacs to avoid
this problem, but for now just fail on short timeouts and let the caller handle
it.