Basic continuous dictation system
Allows symbols and special keypresses to be spoken during formatted dictation.
They'll be inserted differently depending on the formatter - some will be
auto-formatted, some will be inserted straight.

A side effect of this is that multiple formatted phrases can be broken up
without the need for another formatting prefix. This makes dictation much faster
for languages with uniform name formatting (e.g. names are usually
space-separated in lisp).