Rework formatters to use chunked phrases
This allows formatted text to be chained with punctuation (and other
keypresses), making for a much smoother dictation experience.

Also move formatter actions out of the base `text` module and into a new
`formatting` module.

Bindings using simple dictation and the old formatting system have not been
updated yet, with the exception of basic formatters. They may be broken until
they are updated.